Remove dependency from on zio-config from `tools`
