+++ build.sbt
@@ -170,29 +170,24 @@ lazy val core = project
   .settings(enableMimaSettingsJVM)
   .settings(
     testFrameworks := Seq(new TestFramework("zio.test.sbt.ZTestFramework")),
-    libraryDependencies ++= {
-      if (scalaVersion.value == scala212) {
-        Seq("org.scala-lang.modules" %% "scala-collection-compat" % "2.11.0")
-      } else {
-        Seq()
-      }
-    } ++
+    libraryDependencies ++=
       Seq(
-        "com.lihaoyi"                           %% "fastparse"             % "3.0.2",
-        "dev.zio"                               %% "zio"                   % zioVersion,
-        "dev.zio"                               %% "zio-streams"           % zioVersion,
-        "dev.zio"                               %% "zio-query"             % zqueryVersion,
-        "dev.zio"                               %% "zio-prelude"           % zioPreludeVersion,
-        "dev.zio"                               %% "zio-test"              % zioVersion      % Test,
-        "dev.zio"                               %% "zio-test-sbt"          % zioVersion      % Test,
-        "dev.zio"                               %% "zio-json"              % zioJsonVersion  % Optional,
-        "com.softwaremill.sttp.tapir"           %% "tapir-core"            % tapirVersion    % Optional,
-        "io.circe"                              %% "circe-core"            % circeVersion    % Optional,
-        "io.circe"                              %% "circe-parser"          % circeVersion    % Test,
-        "com.github.plokhotnyuk.jsoniter-scala" %% "jsoniter-scala-core"   % jsoniterVersion % Optional,
-        "com.github.plokhotnyuk.jsoniter-scala" %% "jsoniter-scala-macros" % jsoniterVersion % Provided,
-        "org.playframework"                     %% "play-json"             % playJsonVersion % Optional,
-        "org.apache.commons"                     % "commons-lang3"         % "3.14.0"        % Test
+        "com.lihaoyi"                           %% "fastparse"               % "3.0.2",
+        "org.scala-lang.modules"                %% "scala-collection-compat" % "2.11.0",
+        "dev.zio"                               %% "zio"                     % zioVersion,
+        "dev.zio"                               %% "zio-streams"             % zioVersion,
+        "dev.zio"                               %% "zio-query"               % zqueryVersion,
+        "dev.zio"                               %% "zio-prelude"             % zioPreludeVersion,
+        "dev.zio"                               %% "zio-test"                % zioVersion      % Test,
+        "dev.zio"                               %% "zio-test-sbt"            % zioVersion      % Test,
+        "dev.zio"                               %% "zio-json"                % zioJsonVersion  % Optional,
+        "com.softwaremill.sttp.tapir"           %% "tapir-core"              % tapirVersion    % Optional,
+        "io.circe"                              %% "circe-core"              % circeVersion    % Optional,
+        "io.circe"                              %% "circe-parser"            % circeVersion    % Test,
+        "com.github.plokhotnyuk.jsoniter-scala" %% "jsoniter-scala-core"     % jsoniterVersion % Optional,
+        "com.github.plokhotnyuk.jsoniter-scala" %% "jsoniter-scala-macros"   % jsoniterVersion % Provided,
+        "org.playframework"                     %% "play-json"               % playJsonVersion % Optional,
+        "org.apache.commons"                     % "commons-lang3"           % "3.14.0"        % Test
       )
   )
   .dependsOn(macros)
@@ -221,8 +216,6 @@ lazy val tools = project
       "org.scalameta"                  % "scalafmt-interfaces" % scalafmtVersion,
       "io.get-coursier"                % "interface"           % "1.0.19",
       "com.softwaremill.sttp.client3" %% "zio"                 % sttpVersion,
-      "dev.zio"                       %% "zio-config"          % zioConfigVersion,
-      "dev.zio"                       %% "zio-config-magnolia" % zioConfigVersion,
       "dev.zio"                       %% "zio-test"            % zioVersion     % Test,
       "dev.zio"                       %% "zio-test-sbt"        % zioVersion     % Test,
       "dev.zio"                       %% "zio-json"            % zioJsonVersion % Test
@@ -268,7 +261,9 @@ lazy val codegenSbt = project
     crossScalaVersions := Seq(scala212),
     testFrameworks     := Seq(new TestFramework("zio.test.sbt.ZTestFramework")),
     libraryDependencies ++= Seq(
-      "dev.zio" %% "zio-test-sbt" % zioVersion % Test
+      "dev.zio" %% "zio-config"          % zioConfigVersion,
+      "dev.zio" %% "zio-config-magnolia" % zioConfigVersion,
+      "dev.zio" %% "zio-test-sbt"        % zioVersion % Test
     )
   )
   .enablePlugins(SbtPlugin)
