+++ codegen-sbt/src/test/scala/caliban/codegen/OptionsParserSpec.scala
@@ -1,14 +1,15 @@
-package caliban.tools
+package caliban.codegen
 
+import caliban.tools.Options
 import caliban.tools.Options.Header
-import zio.test._
+import zio.test.*
 
-object OptionsSpec extends ZIOSpecDefault {
+object OptionsParserSpec extends ZIOSpecDefault {
   override def spec =
-    suite("OptionsSpec")(
+    suite("OptionsParserSpec")(
       test("full arguments") {
         val input = List("schema", "output", "--scalafmtPath", "fmtPath", "--headers", "header1:value1,header2:value2")
-        Options.fromArgs(input).map { result =>
+        OptionsParser.fromArgs(input).map { result =>
           assertTrue(
             result == Some(
               Options(
@@ -36,7 +37,7 @@ object OptionsSpec extends ZIOSpecDefault {
       },
       test("full arguments (--headers option first)") {
         val input = List("schema", "output", "--headers", "header1:value1,header2:value2", "--scalafmtPath", "fmtPath")
-        Options.fromArgs(input).map { result =>
+        OptionsParser.fromArgs(input).map { result =>
           assertTrue(
             result == Some(
               Options(
@@ -64,7 +65,7 @@ object OptionsSpec extends ZIOSpecDefault {
       },
       test("minimum arguments") {
         val input = List("schema", "output")
-        Options.fromArgs(input).map { result =>
+        OptionsParser.fromArgs(input).map { result =>
           assertTrue(
             result ==
               Some(
@@ -93,25 +94,25 @@ object OptionsSpec extends ZIOSpecDefault {
       },
       test("not enough arguments") {
         val input = List("schema")
-        Options.fromArgs(input).map { result =>
+        OptionsParser.fromArgs(input).map { result =>
           assertTrue(result == None)
         }
       },
       test("--scalafmtPath value missing") {
         val input = List("schema", "output", "--scalafmtPath", "--headers", "header1:value1,header2:value2")
-        Options.fromArgs(input).map { result =>
+        OptionsParser.fromArgs(input).map { result =>
           assertTrue(result == None)
         }
       },
       test("empty list") {
         val input = Nil
-        Options.fromArgs(input).map { result =>
+        OptionsParser.fromArgs(input).map { result =>
           assertTrue(result == None)
         }
       },
       test("provide package name") {
         val input = List("schema", "output", "--packageName", "com.github.ghostdogpr")
-        Options.fromArgs(input).map { result =>
+        OptionsParser.fromArgs(input).map { result =>
           assertTrue(
             result ==
               Some(
@@ -140,7 +141,7 @@ object OptionsSpec extends ZIOSpecDefault {
       },
       test("provide client name") {
         val input = List("schema", "output", "--clientName", "GraphqlClient.scala")
-        Options.fromArgs(input).map { result =>
+        OptionsParser.fromArgs(input).map { result =>
           assertTrue(
             result ==
               Some(
@@ -169,7 +170,7 @@ object OptionsSpec extends ZIOSpecDefault {
       },
       test("provide effect") {
         val input = List("schema", "output", "--effect", "cats.effect.IO")
-        Options.fromArgs(input).map { result =>
+        OptionsParser.fromArgs(input).map { result =>
           assertTrue(
             result ==
               Some(
@@ -198,7 +199,7 @@ object OptionsSpec extends ZIOSpecDefault {
       },
       test("provide genView") {
         val input = List("schema", "output", "--genView", "true")
-        Options.fromArgs(input).map { result =>
+        OptionsParser.fromArgs(input).map { result =>
           assertTrue(
             result ==
               Some(
@@ -227,7 +228,7 @@ object OptionsSpec extends ZIOSpecDefault {
       },
       test("provide extensibleEnums") {
         val input = List("schema", "output", "--extensibleEnums", "true")
-        Options.fromArgs(input).map { result =>
+        OptionsParser.fromArgs(input).map { result =>
           assertTrue(
             result ==
               Some(
@@ -256,7 +257,7 @@ object OptionsSpec extends ZIOSpecDefault {
       },
       test("provide scalarMappings") {
         val input = List("schema", "output", "--scalarMappings", "Long:scala.Long")
-        Options.fromArgs(input).map { result =>
+        OptionsParser.fromArgs(input).map { result =>
           assertTrue(
             result ==
               Some(
@@ -285,7 +286,7 @@ object OptionsSpec extends ZIOSpecDefault {
       },
       test("provide imports") {
         val input = List("schema", "output", "--imports", "a.b.Clazz,b.c._")
-        Options.fromArgs(input).map { result =>
+        OptionsParser.fromArgs(input).map { result =>
           assertTrue(
             result ==
               Some(
@@ -314,7 +315,7 @@ object OptionsSpec extends ZIOSpecDefault {
       },
       test("provide abstractEffectType") {
         val input = List("schema", "output", "--effect", "F", "--abstractEffectType", "true")
-        Options.fromArgs(input).map { result =>
+        OptionsParser.fromArgs(input).map { result =>
           assertTrue(
             result ==
               Some(
@@ -343,7 +344,7 @@ object OptionsSpec extends ZIOSpecDefault {
       },
       test("provide preserveInputNames") {
         val input = List("schema", "output", "--preserveInputNames", "true")
-        Options.fromArgs(input).map { result =>
+        OptionsParser.fromArgs(input).map { result =>
           assertTrue(
             result ==
               Some(
@@ -372,7 +373,7 @@ object OptionsSpec extends ZIOSpecDefault {
       },
       test("header with a colon in the value") {
         val input = List("schema", "output", "--scalafmtPath", "fmtPath", "--headers", "aaa:bbb:ccc")
-        Options.fromArgs(input).map { result =>
+        OptionsParser.fromArgs(input).map { result =>
           assertTrue(
             result ==
               Some(
